Implement content() method than can decouple element begin and end.
In the previous version, only lambda element bodies were supported,
which meant there was no way to being a tag, do something else, then
end the tag later.

This begin, do something else, end pattern is required for legacy JSP
taglib implementations, as one example.

This new release also supports a new content() method, which is
AutoCloseable for try-with-resources or may be closed with __().  When
closed with __(), the parent content model is returned, thus allowing
fluent-style method chaining both into and back out of element content.